Collection: move ArrayCollection to collection data file
